requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
numpy
scipy
matplotlib
scikit-image
Pillow
tqdm
kagglehub